Refactor e2e test utils so that we can use them in other tests
